# Conflicts:
#	Source/Tam/Maestro/Services/Broadcast/Services.Broadcast/ApplicationServices/AffidavitPreprocessingService.cs
#	Source/Tam/Maestro/UI/BroadcastComposerWebUI/src/app/components/planning/ProposalForm/MarketGroupSelector/index.js

+++ Source/Solutions/WWTVData.Service/WWTVDataFile.cs
@@ -112,11 +112,11 @@ namespace WWTVData.Service
                 var response = service.ProcessFiles(filesFound.ToList(), ServiceName);
                 response.ForEach(r =>
                 {
-                    if (r.Status == (int) AffidaviteFileProcessingStatus.Valid)
+                    if (r.Status == AffidaviteFileProcessingStatus.Valid)
                     {
                         File.Delete(r.FilePath);
                         filesProcessed++;
-                    } else if (r.Status == (int) AffidaviteFileProcessingStatus.Valid)
+                    } else if (r.Status == AffidaviteFileProcessingStatus.Valid)
                         filesFailed++;
                 });
                 _LastRun = DateTime.Now;
